test(test_description_t): temp nagfor workaround
This commit uses the NAG compiler's predefined NAGFOR macro to
remove two lines in test_description_test.F90 to work around a
bug in the nagfor compiler. The compiler bug has been isolated and
reported and the associated macro conditionals can be removed once
a patched version of the compiler is available. In the interim,
simply substituting character variables for string_t arguments in
the test's contains_text check works around the issue.