fix: work around gfortran bugs
This commit works around apparent gfortran compiler bugs that
blocked building the parent commit with gfortran 14 and 15.
With this commit's workarounds, gfortran 15.0.1 20250119
(experimental) is able to build Julienne.  However, the
newest code in vector_test_description_test_m crashes with a
segmentation fault even with assertions on (-DASSERTIONS=1).

Once this branch is merged into main, we will likely have to drop
support for gfortran versions less than 15.  Further investigation
is needed to identify and work around the issue that causes a seg
fault when compiling with gfortran 15.