feat(test_description_t): automate skipping tests
The test_description_t and vector_test_description_t constructors'
diagnosis_function and vector_diagnosic_function arguments are now
optional.  When these arguments are not present, no diagnostic
function executes, resulting in test_result_t "run" procedure
reporting the test as skipped.  This features is useful for tests
that crash with specific compilers.  Skipping a test documents that
the a feature is not supported with a given compiler.  For example,
both of the vector_test_description_t tests are currently skipped
in the Julienne repository, indicating that we cannot currently
support use of that type with gfortarn.