feat(test_diagnosis): add comparison operators
This commit adds user-defined operators and corresponding unit
tests.  The operators that support statements of the form

  type(test_diagnosis_t) test_diagnosis
  test_diagnosis = x .approximates. y .within. tolerance

for real and double-precsision operands x, y, and tolerance and

  test_diagnosis = i .equalsExpected. j

for integer operand i and j. The right-hand side expressions above
evaluate to test_diagnostics_t objects with diagnostics_string_
components formed with the assumption that x and i are actual
computed values y and j corresponding expected values, respectively.