feat(string_t): build markdown table from array
This commit adds support for, and an example demonstrating,
function invocations of the form

  integer, parameter :: array(*,*) = reshape([1,2,3,4], [2,2])
  markdown_table(string_t(array), side_borders=.true.)

where the function result contains an 1D string_t array in which
each element represents one row of a Markdown table created from
one row of the 2D integer_array.  The above approach leverages the
elemental attribute of the string_t constructor function and
therefore also supports real, double precision, and complex
values.  The approach also works for string_t arrays, in which
case there is no need to invoke the string_t constructor.