fix(macros): work around preprocessor(s)
The C11 standard stipulates, "If the name of the macro being
replaced is found during this scan of the replacement list...,
it is not replaced."  LLVM's C preprocessor appears to honor this
stipulation.  Gfortran's and nagfor's preprocessors appear to be
non-conforming in that they allow recursive expansion.  This
commit eliminates the possibility of recursion by appending an
underscore to the `call_julienne_assert` generic interface so
that it no longer matches the corresponding function-like macro
name.  This leads to a slight and unfortunate distinction: the
name that appears in the only clause of a `use` statement differs
by one character (the trailing underscore) from the name used at
the point of invocation of the macro.