BREAKING: rm vector_test_description_{m,s}
This commit prepares for moving the test-filtering logic to one
central location in test_description_s rather than having versions
of the logic distributed across all tests.  Because the
deprecated vector_test_description_t is convoluted due to
preprocessor  directives, extracting the filter from that file
would be challenging, which motivates the removal.  This is a
breaking change for any codes that use vector_test_description_t.