fix(get_json_value): work around GCC 14/15 issue
This commit makes the specific string_t binding get_integer_array
public when compiling with gfortran to work around a seg fault that
occurs when this procedure is invoked through the get_json_value
generic binding.  The corresponding unit test is adjusted to use
the specific binding until a minimal reproducer can be developed
and reported, at which point it could make sense to instead skip the
test as a way to document and expose the issue in the test output.