test_description_test_m: Fix spurious failures with gcc
In the absence of -O3, it has been observed with gcc 14.2..15.2
that when pproc is a pointer to proc, the combination
  c_f_procpointer(c_funloc(proc), pproc2)
can produce a pproc2 that differs from pproc by 12 bytes.
However the expression:
  c_f_procpointer(c_funloc(pproc), pproc2)
(passing a procedure pointer to c_funloc rather than the procedure itself)
seems to more reliably produce a pproc2 that matches pproc.

This might be a compiler bug, or possibly intentional behavior.
It's unclear whether this is a violation of the standard.